MISUV-7493: code refactor
